@@ -13,19 +13,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-name: generation_tests_pre_212
-description: Sample of using instant_xml_parser to generate a parser via annotation.
+name: testing
+description: Validating generated parsers.
 version: 1.0.0
 homepage: https://github.com/sarahec/instant_xml_parser
 
 environment:
-  sdk: '>=2.10.0 < 3.0.0'
+  sdk: '>=2.12.0 < 3.0.0'
 
 dependencies:
   logging: ^1.0.0
   xml: ^5.0.2
-  instant_xml_parser: ^0.11.0
-  ixp_runtime: ^0.11.0
+  instant_xml_parser: ^1.0.0
+  ixp_runtime: ^1.0.0
 
 dev_dependencies:
   build_runner: ^1.0.0
@@ -34,7 +34,9 @@ dev_dependencies:
 
 # Added to make developing new features easier. Don't use in your own code.
 dependency_overrides:
-  # instant_xml_parser:
-  #   path: ../instant_xml_parser
+  instant_xml_parser:
+    path: ../instant_xml_parser
+  ixp_runtime:
+    path: ../ixp_runtime
   logging: ^1.0.0
   xml: ^5.0.2
